+++ platforms/baidu/wrapper/engine/Loader.js
@@ -1,0 +1,16 @@
+    // baidu can`t use system So need special treatment
+    cc.loader.downloader.loadSubpackage = function(name, completeCallback) {
+        swan.loadSubpackage({
+            name: name,
+            success: function() {
+                cc$System.import('virtual:///prerequisite-imports/' + name).then(function() {
+                    if (completeCallback) { completeCallback(); }
+                }).catch(function(err) {
+                    if (completeCallback) { completeCallback(err); }
+                });
+            },
+            fail: function() {
+                if (completeCallback) { completeCallback(new Error(`Failed to load subpackage ${name}`)); }
+            },
+        });
+    };
